+++ got2d/got2d.vcxproj.filters
@@ -85,5 +85,8 @@
     <ClCompile Include="source\message_win32.cpp">
       <Filter>源文件\input</Filter>
     </ClCompile>
+    <ClCompile Include="source\scene_node.cpp">
+      <Filter>源文件\scene</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
